Respect local system permissions with certificate provisioner (#71)
The CI is not allowed to connect to Tailnet devices as root, therefore
actions to update uhttpd settings via UCI are restricted. It will be
assumed that the required values will be pre-set in UCI, so this CI only
has to restart uhttpd. That permission has been added via sudoers
changes on the target device.